Add modular quantum circuit framework
