+++ src/quantum/__init__.py
@@ -1,0 +1,6 @@
+"""Quantum computing utilities for custom algorithms."""
+
+from .gates import H, X, I, CNOT
+from .circuit import QuantumCircuit
+
+__all__ = ["H", "X", "I", "CNOT", "QuantumCircuit"]
