+++ src/quantum/gates.py
@@ -1,0 +1,11 @@
+"""Basic quantum gates expressed as numpy matrices."""
+
+import numpy as np
+
+H = (1/np.sqrt(2)) * np.array([[1, 1], [1, -1]], dtype=complex)
+X = np.array([[0, 1], [1, 0]], dtype=complex)
+I = np.eye(2, dtype=complex)
+CNOT = np.array([[1, 0, 0, 0],
+                 [0, 1, 0, 0],
+                 [0, 0, 0, 1],
+                 [0, 0, 1, 0]], dtype=complex)
